// src/types/view-decline-cases-type.ts
export interface declineLeadCases {
  lead_capture_id: number;
  name: string;
  email: string;
  phone: string;
  coreAcademicCareerId: number;
  academic_career_description: string;
  coreAcademicProgramId: number;
  academic_program_description: string;
  coreLeadStageId: number;
  leadStageDescription: string;
  coreLeadSubStageId: number;
  current_lead_sub_stage_display_name: string;
  modifiedOn: string | any;
  lead_source_description: string;
  status: string | any;
  created_at: string | any;
  leadOwner: string | any;
}